Add files via upload
This repository contains my finals project for Intuitas Data Science NEET Academy in Trokut, Šibenik. 

Main goal of the project was to create a simple ML model that will predict outcomes of the matches. Data used in the making of machine learning model was scraped from basketball-reference.com via playwright scraper, which was later preprocessed for ML. 

There are 4 notebooks with "WIP" prefix, meaning that's work in progress and needs to be finished.